+++ src/sentry/api/endpoints/organization_projects.py
@@ -1,16 +1,16 @@
-from typing import Any, List, Optional
-
 from django.db.models import Q
 from drf_spectacular.utils import OpenApiExample, extend_schema
 from rest_framework.request import Request
 from rest_framework.response import Response
-from typing_extensions import TypedDict
 
 from sentry.api.base import EnvironmentMixin
 from sentry.api.bases.organization import OrganizationEndpoint
 from sentry.api.paginator import OffsetPaginator
 from sentry.api.serializers import serialize
-from sentry.api.serializers.models.project import ProjectSummarySerializer
+from sentry.api.serializers.models.project import (
+    OrganizationProjectResponseDict,
+    ProjectSummarySerializer,
+)
 from sentry.apidocs.constants import RESPONSE_FORBIDDEN, RESPONSE_NOTFOUND, RESPONSE_UNAUTHORIZED
 from sentry.apidocs.parameters import CURSOR_QUERY_PARAM, GLOBAL_PARAMS
 from sentry.apidocs.utils import inline_sentry_response_serializer
@@ -18,42 +18,6 @@ from sentry.models import Project, ProjectStatus, Team
 from sentry.search.utils import tokenize_query
 
 ERR_INVALID_STATS_PERIOD = "Invalid stats_period. Valid choices are '', '24h', '14d', and '30d'"
-
-
-class TeamResponseDict(TypedDict):
-    id: str
-    name: str
-    slug: str
-
-
-class EventProcessingDict(TypedDict):
-    symbolicationDegraded: bool
-
-
-class LatestReleaseDict(TypedDict):
-    version: str
-
-
-class OrganizationProjectResponseDict(TypedDict):
-    team: TeamResponseDict
-    teams: List[TeamResponseDict]
-    id: str
-    name: str
-    slug: str
-    isBookmarked: bool
-    isMember: bool
-    hasAccess: bool
-    dateCreated: str
-    environments: Optional[List[str]]
-    eventProcessing: EventProcessingDict
-    features: List[str]
-    firstEvent: Optional[str]
-    firstTransactionEvent: bool
-    hasSessions: bool
-    platform: str
-    platforms: List[str]
-    latestRelease: Optional[LatestReleaseDict]
-    latestDeploys: Optional[Any]
 
 
 @extend_schema(tags=["Organizations"])
@@ -74,7 +38,7 @@ class OrganizationProjectsEndpoint(OrganizationEndpoint, EnvironmentMixin):
         },
         examples=[
             OpenApiExample(
-                "Successful Response",
+                "Success",
                 value={
                     "dateCreated": "2018-11-06T21:19:58.536Z",
                     "firstEvent": None,
